Better Handling of errors with returning 304, and 401

+++ rchip/views.py
@@ -4,18 +4,19 @@ from schedule.models import episode_data,tv_shows,user_tv_shows
 from django.contrib.auth.models import User
 from django.contrib.auth import logout,login,authenticate
 from django.contrib.auth.forms import UserCreationForm
+from django.http import HttpResponse, HttpResponseNotModified,HttpResponseForbidden
 from django.views.decorators.csrf import csrf_exempt
 import os,time
-from datetime import datetime
+from datetime import datetime,timedelta
 
 @csrf_exempt
 def json_get_daemons(request):
     id = get_id(request)
     if id is not None:
-        daemons = daemon_register.objects.all().filter(user=User.objects.get(id=id))
+        daemons = daemon_register.objects.all().filter(user__id=id)
         return JSONResponse(daemons.values('hostname'),Extra={"success":True})
     else :
-        return JSONResponse(None,Extra={"success":False,"message":"Not Authorized"})
+        return HttpResponseForbidden()
 
 @csrf_exempt
 def json_get_video_path(request):
@@ -23,12 +24,12 @@ def json_get_video_path(request):
     if id is not None:
         host = request.GET['host']
         if host!=None:
-            path = daemon_register.objects.all().filter(hostname=host,user=User.objects.get(id=id))
+            path = daemon_register.objects.all().filter(hostname=host,user__id=id)
         else:
             path = daemon_register.objects.all()
         return JSONResponse(path.values('path_to_root'),Extra={"success":True})
     else :
-        return JSONResponse(None,Extra={"success":False,"message":"Not Authorized"})
+        return HttpResponseForbidden()
 
 @csrf_exempt
 def json_register_daemon(request):
@@ -43,14 +44,15 @@ def json_register_daemon(request):
                 daemon.path_to_root=video_path
                 daemon.user = User.objects.get(id=id)
             except:
-                daemon = daemon_register(hostname=hostname_in,path_to_root=video_path,user=User.objects.get(id=id))
+                new_user = User.objects.get(id=id)
+                daemon = daemon_register(hostname=hostname_in,path_to_root=video_path,user__id=new_user)
             daemon.save()
             response['success']=True
         else:
             response['success']=False
         return JSONResponse(response,Extra={"success":True})
     else:
-        return JSONResponse(None,Extra={"success":False,"message":"Not Authorized"})
+        return HttpResponseForbidden()
 
 @csrf_exempt
 def json_send_command(request):
@@ -62,14 +64,15 @@ def json_send_command(request):
         source_hostname_in = request.GET['source_hostname']
         destination_hostname_in = request.GET['destination_hostname']
         if (command_in != None):
-            com_que = command_queue(command=command_in,command_text=command_text_in,source_hostname=source_hostname_in,destination_hostname=destination_hostname_in,user=User.objects.get(id=id))
+            new_user = User.objects.get(id=id)
+            com_que = command_queue(command=command_in,command_text=command_text_in,source_hostname=source_hostname_in,destination_hostname=destination_hostname_in,user=new_user)
             com_que.save()
             response['success']=True
         else:
             response['success']=False
         return JSONResponse(response,Extra={"success":True})
     else:
-        return JSONResponse(None,Extra={"success":False,"message":"Not Authorized"})
+        return HttpResponseForbidden()
 
 
 @csrf_exempt
@@ -80,26 +83,27 @@ def json_register_remote_device(request):
         device_name_in = request.GET['device_name']
         state_in = request.GET['state'] in ['true','TRUE','True','T','t','1']
         if (device_name_in!=None):
+            new_user = User.objects.get(id=id)
             try:
                 rem_dev = remote_devices.objects.get(devices_name=device_name_in)
                 rem_dev.active = state_in
-                rem_dev.user = User.objects.get(id=id)
+                rem_dev.user = new_user
             except:
-                rem_dev = remote_devices(devices_name=device_name_in,active=state_in,user=User.objects.get(id=id))
+                rem_dev = remote_devices(devices_name=device_name_in,active=state_in,user=new_user)
             rem_dev.save()
             response['success']=True
         else:
             response['success']=False
         return JSONResponse(response,Extra={"success":True})
     else :
-        return JSONResponse(None,Extra={"success":False,"message":"Not Authorized"})
+        return HttpResponseForbidden()
 
 @csrf_exempt
 def json_get_remote_device(request):
     id = get_id(request)
     if id is not None:
         try:
-            return JSONResponse(remote_devices.objects.all().filter(active=True,user=User.objects.get(id=id)).values('devices_name'),Extra={"success":True})
+            return JSONResponse(remote_devices.objects.all().filter(active=True,user__id=id).values('devices_name'),Extra={"success":True})
         except:
             return JSONResponse(None)
 
@@ -109,11 +113,11 @@ def json_get_song_info(request):
     if id is not None:
         host = request.GET['host']
         if host!=None:
-            mus = music_info.objects.all().filter(destination_hostname=host,user=User.objects.get(id=id))
+            mus = music_info.objects.all().filter(destination_hostname=host,user__id=id)
             return JSONResponse(mus.values('artist','album','song','total_time','is_playing'),Extra={"success":True})
         return JSONResponse(Extra={"success":False})
     else :
-        return JSONResponse(None,Extra={"success":False,"message":"Not Authorized"})
+        return HttpResponseForbidden()
 
 @csrf_exempt
 def json_set_song_info(request):
@@ -125,7 +129,8 @@ def json_set_song_info(request):
         t_time = request.GET['total_time']
         is_playing = True if request.GET['is_playing'] == 1 else False
         d_hostname = request.GET['dest_hostname']
-        obj, created = music_info.objects.get_or_create(destination_hostname=d_hostname, defaults={'artist':artist,'album':album,'song':song,'total_time':t_time,'is_playing':is_playing},user=User.objects.get(id=id))
+        new_user = User.objects.get(id=id)
+        obj, created = music_info.objects.get_or_create(destination_hostname=d_hostname, defaults={'artist':artist,'album':album,'song':song,'total_time':t_time,'is_playing':is_playing},user=new_user)
         if not created:
             obj.artist = artist
             obj.album = artist
@@ -136,7 +141,7 @@ def json_set_song_info(request):
         obj.save()
         return JSONResponse(None,Extra={"success":True})
     else :
-        return JSONResponse(None,Extra={"success":False,"message":"Not Authorized"})
+        return nHttpResponseForbidden()
 
 @csrf_exempt
 def json_get_command(request):
@@ -144,13 +149,15 @@ def json_get_command(request):
     if id is not None:
         host = request.GET['host']
         if host!=None:
-            command = command_queue.objects.all().filter(destination_hostname=host,user=User.objects.get(id=id))
+            command = command_queue.objects.all().filter(destination_hostname=host,user__id=id)
+            if not command.count():
+                return HttpResponseNotModified()
             retval = JSONResponse(command.values('command','command_text'),Extra={"success":True})
             command.delete()
             return retval
         return JSONResponse(None,Extra={"success":True})
     else :
-        return JSONResponse(None,Extra={"success":False,"message":"Not Authorized"})
+        return HttpResponseForbidden()
 
 @csrf_exempt
 def json_show_downloaded(request):
@@ -169,7 +176,8 @@ def json_show_downloaded(request):
         ticker = "%s Downloaded|%s|%s" %(file_name,file_path,file_name)
     #TODO: Here we need to only get remote devices in which show_view is checled
     for host in remote_devices.objects.all().filter(active=True):
-        cmd = command_queue(command='TMSG',command_text=ticker,source_hostname='FILE_SERVER',destination_hostname=host.devices_name,user=User.objects.get(id=host.user_id))
+        new_user = User.objects.get(id=host.user_id)
+        cmd = command_queue(command='TMSG',command_text=ticker,source_hostname='FILE_SERVER',destination_hostname=host.devices_name,user=new_user)
         cmd.save()
     return JSONResponse(None,Extra={"success":True})
 
@@ -260,10 +268,10 @@ def json_get_upcoming_shows(request):
     now = datetime.now()
     id=get_id(request)
     if id != None:
-        eps_list = episode_data.objects.filter(show__user_tv_shows__user=User.objects.get(id=id),air_date__gte=now)
+        eps_list = episode_data.objects.filter(show__user_tv_shows__user__id=id,air_date__gte=now)
         return JSONResponse(eps_list.values('show__name','air_date','eps_name','eps_number','show__air_time'),Extra={"success":True})
     else:
-        return JSONResponse(None,Extra={"success":False,"message":"Not Authorized"})
+        return HttpResponseForbidden()
 
 
 def get_id(request):
